+++ app/src/main/java/exh/ui/metadata/MetadataViewViewModel.kt
@@ -1,13 +1,16 @@
 package exh.ui.metadata
 
-import cafe.adriel.voyager.core.model.StateScreenModel
-import cafe.adriel.voyager.core.model.screenModelScope
+import androidx.lifecycle.viewModelScope
+import androidx.lifecycle.viewmodel.CreationExtras
+import androidx.lifecycle.viewmodel.initializer
+import androidx.lifecycle.viewmodel.viewModelFactory
 import eu.kanade.domain.ui.UiPreferences
 import eu.kanade.tachiyomi.source.online.MetadataSource
 import exh.metadata.metadata.RaisedSearchMetadata
 import exh.source.getMainSource
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import mihon.core.viewmodel.StateViewModel
 import tachiyomi.core.common.util.lang.launchIO
 import tachiyomi.domain.manga.interactor.GetFlatMetadataById
 import tachiyomi.domain.manga.interactor.GetManga
@@ -16,13 +19,13 @@ import tachiyomi.domain.source.service.SourceManager
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
 
-class MetadataViewScreenModel(
+class MetadataViewViewModel(
     val mangaId: Long,
     val sourceId: Long,
     private val getFlatMetadataById: GetFlatMetadataById = Injekt.get(),
     private val sourceManager: SourceManager = Injekt.get(),
     private val getManga: GetManga = Injekt.get(),
-) : StateScreenModel<MetadataViewState>(MetadataViewState.Loading) {
+) : StateViewModel<MetadataViewState>(MetadataViewState.Loading) {
 
     private val uiPreferences = Injekt.get<UiPreferences>()
     val themeCoverBased = uiPreferences.themeCoverBased.get()
@@ -31,11 +34,11 @@ class MetadataViewScreenModel(
     val manga = _manga.asStateFlow()
 
     init {
-        screenModelScope.launchIO {
+        viewModelScope.launchIO {
             _manga.value = getManga.await(mangaId)
         }
 
-        screenModelScope.launchIO {
+        viewModelScope.launchIO {
             val metadataSource = sourceManager.get(sourceId)?.getMainSource<MetadataSource<*, *>>()
             if (metadataSource == null) {
                 mutableState.value = MetadataViewState.SourceNotFound
@@ -45,6 +48,20 @@ class MetadataViewScreenModel(
             mutableState.value = when (val flatMetadata = getFlatMetadataById.await(mangaId)) {
                 null -> MetadataViewState.MetadataNotFound
                 else -> MetadataViewState.Success(flatMetadata.raise(metadataSource.metaClass))
+            }
+        }
+    }
+
+    companion object {
+        val MANGA_ID_KEY = CreationExtras.Key<Long>()
+        val SOURCE_ID_KEY = CreationExtras.Key<Long>()
+
+        val Factory = viewModelFactory {
+            initializer {
+                MetadataViewViewModel(
+                    mangaId = get(MANGA_ID_KEY)!!,
+                    sourceId = get(SOURCE_ID_KEY)!!,
+                )
             }
         }
     }